Create a new student_tabs.php class used to render the tabs for student_evalutions.php, student_certifications.php and student_evalutations.php Added in a header showing the current user too
The eval tabs now shows the grade input of the observers (if any). Still a work in
progress. I will add the autoeval here too

The cases are shown on the List tab

For the Grading UI the charts are showing again, needs some love but works.

Cleaned out old build JS files