Multiple improvements
CompetVet instances can use a custom evaluation criteria
Custom criteria can be configured in a new criteria tab
Global criteria are locked, in a future feature this will depend on any grades given to users
Fixed issue with Grade button not appearing on a view pages for a user
When adding a case for a user using the modal form, reload page when form is done so the new case is rendered
Fix issue with the default level not set when a user declares their confidence in an evaluation criterium
Check if the list criteria have been updated before using stored formdata for a user. If the criteria have changed this automatically invalidates the formdata.
Fix issue where a 0 grade does not show in the grading app
Fix JS issues on settings pages because of the criteria manager
Allow HTML in the comments when format_text add stuff from filters etc
Add the context object to the base::get_button() to check the permission first, this is used so the grade button does not show up for students
Only show the debugging JSON if $CFG->debugdisplay = true
add the situationid to the competvet_grid table
Remove duplicate button for add case on list tab