Fix #562: Multiple fixes to the table and check if a grade == 50
* When a grade is 50, it means that the observer has not decided, so we
remove it from the computation of the average (grading page)
* There were issues with the displayed tables and the graph because as sometimes
columns of the result table were not aligned and some observation were missing
* The changes introduced here was make a distinction between observations and autoeval
* The tables header are now built independently (before we were counting the number
of column not distinctively and empty rows in eval were not removed if there was an eval
in autoeval for example).
* Empty cells are now displayed correctly (some were sometimes missing)
* Fix missing observations (this was due to the fact that if the first row had
less columns that the other, we did not output them).